chore(types): split interfaces into separate files for better structure
