+++ src/types/experience.ts
@@ -1,0 +1,8 @@
+export interface IExperience {
+  id: string;
+  companyOrSchool: string;
+  title: string;
+  date: string;
+  desc: string;
+  isEducation: false;
+}
