+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./project";
+export * from "./experience";
+export * from "./skill";
