+++ src/types/link.ts
@@ -1,0 +1,4 @@
+export interface ILink {
+  label: string;
+  url: string;
+}
