+++ src/types/skill.ts
@@ -1,0 +1,6 @@
+export interface ISkill {
+  id: string;
+  src: string;
+  alt: string;
+  skill: string;
+}
